Fixing calls to log/add.
The Logger class in Rails 3 supports add/log, but the method requires that you
pass in the severity. There are also methods that infer the severity in
the name (i.e. WARN(message), INFO(message) ).

The BufferedLogger class (used pre Rails 3) does not the log method (an
alias to add) nor the named log-level convenience methods. So, you have
to use 'add' and pass the severity.

Enjoy!