Add the ability to select a language when multiple syntax blocks occur together.
Blocks of code syntax must occur next to each other with no separation.
Javascript will work out which blocks are in which language and add to a menu.
By default the first language is selected for each group.
Selecting a language will change all blocks on the page with the same language.